Move ThemeType (and associated imports) to relative imports instead of absolute
