+++ src/shared-components/loadingSpinner/index.tsx
@@ -1,9 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { ThemeColors } from 'src/constants/themes/types';
 import { useTheme } from 'emotion-theming';
 
 import { LoadingSpinnerContainer, Overlay, Dot } from './style';
+import { ThemeColors } from '../../constants';
 
 type LoadingSpinnerProps = {
   /**
