+++ src/shared-components/progressBar/index.tsx
@@ -1,9 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { useTheme } from 'emotion-theming';
-import { ThemeColors } from 'src/constants/themes/types';
 
-import { PROGRESS_BAR_STATUS } from '../../constants';
+import { PROGRESS_BAR_STATUS, ThemeColors } from '../../constants';
 import { OuterContainer, InnerBar } from './style';
 
 export type ProgressBarStatusType = valueof<typeof PROGRESS_BAR_STATUS>;
