+++ src/types/@emotion/styled/index.ts
@@ -1,5 +1,4 @@
-import { ThemeType } from 'src/constants/themes/types';
-
+import { ThemeType } from '../../../constants/themes/types';
 /**
  * We rely on TypeScript path aliasing in order to decorate our `@emotion/styled` import
  * with built-in type information, so for this one import we need to use relative imports
